+++ src/rs_surreal/resolve.rs
@@ -59,10 +59,10 @@ pub fn check_unit_compatible(unit_a: &str, unit_b: &str) -> bool {
             .unwrap_or(false)
 }
 
-pub const INTERNAL_PDMS_EXPRESS: [&'static str; 28] = [
+pub const INTERNAL_PDMS_EXPRESS: [&'static str; 29] = [
     "MAX", "MIN", "COS", "SIN", "LOG", "ABS", "POW", "SQR", "SQRT", "NOT", "AND", "OR", "ATAN",
-    "ACOS", "ATAN2", "ASIN", "INT", "OF", "MOD", "NEGATE", "SUM", "TANF", "TAN", "TIMES", "MULT",
-    "DIV", "ADD", "MINUS",
+    "ACOS", "ATAN2", "ASIN", "INT", "OF", "MOD", "NEGATE", "SUM", "DIFFERENCE", "TANF", "TAN",
+    "TIMES", "MULT", "DIV", "ADD", "MINUS",
 ];
 
 /// 元件库表达式相关的参数